% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_bf_precalculation_results.R
\docType{data}
\name{eq_bf_precalculation_results}
\alias{eq_bf_precalculation_results}
\title{Equivalence Interval Bayes Factor Precalculation Results}
\format{
A dataframe with 33210 rows and 9 variables:
\describe{
  \item{iterate}{numeric, unique id of iteration}
  \item{tpr}{numeric, The long-run probability of obtaining a Bayes factor at least as high as the critical threshold favoring superiority, given Delta.}
  \item{delta}{numeric, The expected population effect size.}
  \item{thresh}{numeric, The Bayes factor threshold for inference. Either 3, 6, or 10.}
  \item{prior_scale}{numeric, Scale of the Cauchy prior distribution. Either 1/sqrt(2), 1, or sqrt(2).}
  \item{eq_band}{numeric, The chosen width of the equivalence region.}
  \item{n1}{numeric, The determined sample size per group.}
  \item{tpr_out}{numeric, The TPR associated with the resulting sample sizes.}
  \item{error_message}{character, The error message in case of an error.}
}
}
\usage{
eq_bf_precalculation_results
}
\description{
Dataset that contains precalculated sample sizes with the
equivalence interval Bayes factor method. The
dataset contains the input and output values of these calculations.
}
\section{Remark}{

  If none of the precalculated values suit your sample size
  determination plan, than feel free the calculations with the R package
  by using the \code{\link{ssp_eq_bf}} function.
}

\keyword{datasets}
