% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informed_t_test_functions.R
\name{informed_t_test_functions}
\alias{informed_t_test_functions}
\alias{cdf_normal}
\alias{cdf_t}
\alias{integrand_t}
\alias{posterior_t}
\alias{posterior_normal}
\title{Informed t-test functions}
\usage{
cdf_normal(
  x,
  t,
  n1,
  n2 = NULL,
  independentSamples = FALSE,
  prior.mean,
  prior.variance,
  rel.tol = .Machine$double.eps^0.25
)

cdf_t(
  x,
  t,
  n1,
  n2 = NULL,
  independentSamples = FALSE,
  prior.location,
  prior.scale,
  prior.df,
  rel.tol = .Machine$double.eps^0.25
)

integrand_t(delta, t, n, nu, mu.delta, gamma, kappa)

posterior_t(
  delta,
  t,
  n1,
  n2 = NULL,
  independentSamples = FALSE,
  prior.location,
  prior.scale,
  prior.df,
  rel.tol = .Machine$double.eps^0.25
)

posterior_normal(
  delta,
  t,
  n1,
  n2 = NULL,
  independentSamples = FALSE,
  prior.mean,
  prior.variance
)
}
\description{
Functions created to calculate Bayes factor with informed priors
for t-tests. The functions were created for the
Gronau, Q. F., Ly, A., & Wagenmakers, E. J. (2019).
Informed Bayesian t-tests. \emph{The American Statistician.} paper.
The functions are retrieved from \href{https://osf.io/bsp6z/}{OSF}.
}
