% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_apriori_precision.R
\name{mod_apriori_precision_ui}
\alias{mod_apriori_precision_ui}
\alias{mod_apriori_precision_server}
\title{mod_apriori_precision_ui and mod_apriori_precision_server}
\usage{
mod_apriori_precision_ui(id)

mod_apriori_precision_server(
  id,
  activate_menu,
  method_menu,
  activate_question,
  method_question
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
