% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_article.R
\name{mod_article_ui}
\alias{mod_article_ui}
\alias{mod_article_server}
\title{mod_article_ui and mod_article_server}
\usage{
mod_article_ui(id)

mod_article_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
