% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_article_module.R
\name{mod_article_module_ui}
\alias{mod_article_module_ui}
\alias{mod_article_module_server}
\title{mod_article_module_ui and mod_article_module_server}
\usage{
mod_article_module_ui(id)

mod_article_module_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
