% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bf_threshold.R
\name{mod_bf_threshold_ui}
\alias{mod_bf_threshold_ui}
\alias{mod_bf_threshold_server}
\title{mod_bf_threshold_ui and mod_bf_threshold_server}
\usage{
mod_bf_threshold_ui(id)

mod_bf_threshold_server(
  id,
  activate_menu,
  method_menu,
  activate_question,
  method_question
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
