% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_flowchart.R
\name{mod_flowchart_ui}
\alias{mod_flowchart_ui}
\alias{mod_flowchart_server}
\title{mod_flowchart_ui and mod_flowchart_server}
\usage{
mod_flowchart_ui(id)

mod_flowchart_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
