% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_interval_equiv_bf.R
\name{mod_interval_equiv_bf_ui}
\alias{mod_interval_equiv_bf_ui}
\alias{mod_interval_equiv_bf_server}
\title{mod_interval_equiv_bf_ui and mod_interval_equiv_bf_server}
\usage{
mod_interval_equiv_bf_ui(id)

mod_interval_equiv_bf_server(
  id,
  activate_menu,
  method_menu,
  activate_question,
  method_question
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
