% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_power_curve.R
\name{mod_power_curve_ui}
\alias{mod_power_curve_ui}
\alias{mod_power_curve_server}
\title{mod_power_curve_ui and mod_power_curve_server}
\usage{
mod_power_curve_ui(id)

mod_power_curve_server(
  id,
  activate_menu,
  method_menu,
  activate_question,
  method_question
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
