% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rope.R
\name{mod_rope_ui}
\alias{mod_rope_ui}
\alias{mod_rope_server}
\title{mod_rope_ui and mod_rope_server}
\usage{
mod_rope_ui(id)

mod_rope_server(
  id,
  activate_menu,
  method_menu,
  activate_question,
  method_question
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
