% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ssp_power_curve.R
\name{mod_ssp_power_curve_ui}
\alias{mod_ssp_power_curve_ui}
\alias{mod_ssp_power_curve_server}
\title{mod_ssp_power_curve_ui and mod_ssp_power_curve_server}
\usage{
mod_ssp_power_curve_ui(id)

mod_ssp_power_curve_server(id)
}
\description{
A shiny Module.
}
\keyword{internal}
