% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ssp_power_traditional.R
\name{mod_ssp_power_traditional_ui}
\alias{mod_ssp_power_traditional_ui}
\alias{mod_ssp_power_traditional_server}
\title{mod_ssp_power_traditional_ui and mod_ssp_power_traditional_server}
\usage{
mod_ssp_power_traditional_ui(id)

mod_ssp_power_traditional_server(id)
}
\description{
A shiny Module.
}
\keyword{internal}
