% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ssp_rope_anova.R
\name{mod_ssp_rope_anova_ui}
\alias{mod_ssp_rope_anova_ui}
\alias{mod_ssp_rope_anova_server}
\title{mod_ssp_rope_anova_ui and mod_ssp_rope_anova_server}
\usage{
mod_ssp_rope_anova_ui(id)

mod_ssp_rope_anova_server(id)
}
\description{
A shiny Module.
}
\keyword{internal}
