% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_traditional_power.R
\name{mod_traditional_power_ui}
\alias{mod_traditional_power_ui}
\alias{mod_traditional_power_server}
\title{mod_traditional_power_ui and mod_traditional_power_server}
\usage{
mod_traditional_power_ui(id)

mod_traditional_power_server(
  id,
  activate_menu,
  method_menu,
  activate_question,
  method_question
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
