% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_power_curve.R
\name{plot_power_curve}
\alias{plot_power_curve}
\title{Function to plot the results of the power curve calculation}
\usage{
plot_power_curve(delta, n1, animated = FALSE)
}
\arguments{
\item{delta}{Numeric. The range of effect sizes.}

\item{n1}{Numeric. The determined sample sizes for each `delta`.}

\item{animated}{Logical. if TRUE the output plot is animated.}
}
\value{
The function returns a `ggplot2` or a `plotly` object.
}
\description{
The function plots the power curve based on the results
of the \code{\link{ssp_power_curve}} function or the \code{\link{ssp_bf_predetermined}} function.
}
\examples{
\dontrun{
# Determine the sample sizes for each delta
curve_data <- SampleSizePlanner::ssp_power_curve(
tpr = 0.8,
delta = seq(0.1, 0.9, 0.01),
max_n = 5000)
# Plot the power curve
SampleSizePlanner::plot_power_curve(delta = curve_data$delta, n1 = curve_data$n1, animated = FALSE)
}
}
