% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope_anova_data.R
\docType{data}
\name{rope_anova_data}
\alias{rope_anova_data}
\title{Region of Practical Equivalence method ANOVA Precalculation Results}
\format{
A dataframe with 5550 rows and 17 variables:
\describe{
  \item{batch_id}{numeric, unique id of batch in which the calculation was run parallel}
  \item{calculation_id}{numeric, unique id of calculation within the batch}
  \item{n1}{numeric, The determined sample size per group.}
  \item{npower}{numeric, The TPR associated with the resulting sample sizes.}
  \item{tpr_out}{numeric, The long-run probability of obtaining a Bayes factor at least as high as the critical threshold favoring superiority, given Delta.}
  \item{tpr_in}{numeric, The long-run probability of obtaining a Bayes factor at least as high as the critical threshold favoring superiority, given Delta.}
  \item{prior_scale}{numeric, Scale of the Cauchy prior distribution. Either 1/sqrt(2), 1, or sqrt(2).}
  \item{eq_band}{numeric, The chosen ROPE interval.}
  \item{error_message}{character, The error message in case of an error.}
  \item{result_not_null}{logical, Whether the precalculation led to an error or not}
}
}
\usage{
rope_anova_data
}
\description{
Dataset that contains precalculated sample sizes with the
ROPE method for 2*2 designes. The dataset contains the input and output values of these calculations.
}
\section{Remark}{

  If none of the precalculated values suit your sample size
  determination plan, than feel free the calculations with the R package
  by using the \code{\link{ssp_rope_anova}} function.
}

\keyword{datasets}
