% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rope_precalculation_results.R
\docType{data}
\name{rope_precalculation_results}
\alias{rope_precalculation_results}
\title{Region of Practical Equivalence method Precalculation Results}
\format{
A dataframe with 11070 rows and 8 variables:
\describe{
  \item{iterate}{numeric, unique id of iteration}
  \item{n1}{numeric, The determined sample size per group.}
  \item{npower}{numeric, The TPR associated with the resulting sample sizes.}
  \item{tpr}{numeric, The long-run probability of obtaining a Bayes factor at least as high as the critical threshold favoring superiority, given Delta.}
  \item{prior_scale}{numeric, Scale of the Cauchy prior distribution. Either 1/sqrt(2), 1, or sqrt(2).}
  \item{delta}{numeric, The expected population effect size.}
  \item{eq_band}{numeric, The chosen ROPE interval.}
  \item{error_message}{character, The error message in case of an error.}
}
}
\usage{
rope_precalculation_results
}
\description{
Dataset that contains precalculated sample sizes with the
ROPE method. The dataset contains the input and output values of these calculations.
}
\section{Remark}{

  If none of the precalculated values suit your sample size
  determination plan, than feel free the calculations with the R package
  by using the \code{\link{ssp_rope}} function.
}

\keyword{datasets}
