% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_aipe.R
\name{ssp_aipe}
\alias{ssp_aipe}
\title{Determine sample size with Accuracy In Parameter Estimation (AIPE) method}
\usage{
ssp_aipe(
  delta,
  confidence_level,
  width,
  which_width = "Full",
  certainty = NULL,
  ...
)
}
\arguments{
\item{delta}{Numeric. The expected population effect size.}

\item{width}{Numeric. The desired width of the confidence interval, given Delta.}

\item{which_width}{Character. The desired width of interest. Either `Full`, `Lower` or `Upper`}

\item{confidence_levell}{Numeric. The desired level of confidence.}

\item{certainity}{Numeric. The desired certainty of the confidence interval width}
}
\value{
The function returns a list of one named element. The determined
sample size for group one and two called `n1`.
}
\description{
A sample size estimation method used for parameter estimation.
The approach aims to find the required sample size, such that
the confidence interval has a certain expected width. The
function is a simple wrapper function around the \code{\link[MBESS]{ss.aipe.smd}}
function.
}
