% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_anova_bf.R
\name{ssp_anova_bf}
\alias{ssp_anova_bf}
\title{Determine sample size with Bayesian Anova}
\usage{
ssp_anova_bf(
  effect,
  mu,
  tpr,
  thresh,
  prior_scale,
  iter,
  max_n = 10001,
  max_bf = 1e+08,
  sigma = 1,
  seed = NULL
)
}
\arguments{
\item{mu}{Numeric. The expected population mean values.}

\item{tpr}{Numeric. The long-run probability of obtaining a Bayes factor at least
as high as the critical threshold favoring superiority, given mu.}

\item{thresh}{Integer. The Bayes factor threshold for inference.}

\item{prior_scale}{Numeric. Scale of the Cauchy prior distribution.}

\item{iter}{Integer. The number of simulations.}

\item{max_n}{Integer. The maximum number of participants per group (all groups are assumed to have equal sample size).}
}
\value{
The function returns a list of four named numeric vectors.
The first `n1` is the range of determined sample sizes for the given design.
The second `tpr_out` is the range of TPRs that were provided as a parameter.
The third `effect` is the type of effect to evaluate, such as "Main Effect 1," "Main Effect 2," or "Interaction Effect."
}
\description{
The present method provides an expected sample size such that
compelling evidence in the form of a Bayes factor can be collected
for a given effect size with a certain long-run probability when
allowing for sequential testing.
}
\examples{
\dontrun{
SampleSizePlanner::ssp_anova_bf(
  effect = "Interaction Effect", tpr = 0.8, max_n = 10001, mu = c(1.5, 1.5, 0, 1), sigma = 2,
  seed = NULL, thresh = 10, prior_scale = 1 / sqrt(2)
 )
}
}
