% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_anova_eq.R
\name{ssp_anova_eq}
\alias{ssp_anova_eq}
\alias{twoway_ANOVA_eq_pwr}
\title{Determine sample size with the Bayesian Interval Equivalence Method}
\usage{
ssp_anova_eq(
  mu,
  effect,
  eq_band,
  tpr,
  thresh,
  prior_scale,
  iter,
  post_iter = 1000,
  sigma = 1,
  prior_location = 0,
  max_n = 10001,
  seed = NULL
)

twoway_ANOVA_eq_pwr(
  n = n1,
  effect = effect,
  eq_band = eq_band,
  iter = iter,
  post_iter = post_iter,
  mu = mu,
  sigma = sigma,
  thresh = thresh,
  prior_scale = prior_scale,
  prior_location = prior_location,
  seed = NULL
)
}
\arguments{
\item{mu}{Numeric. The unstandardized mean of the DV for each group.}

\item{effect}{Character. The effect of interest (main effect A, main effect B).}

\item{eq_band}{Numeric. The margin of the standardized equivalence region.}

\item{tpr}{Numeric. The desired long-run probability of obtaining a Bayes factor higher than Thresh, given the means.}

\item{thresh}{Numeric. The threshold of the Bayes Factor which is fixed to 10 in the ShinyApp.}

\item{prior_scale}{Numeric. The scale of the Cauchy prior which is fixed to 1 / sqrt(2) in the ShinyApp.}

\item{iter}{Numeric. The number of iterations to calculate the TPR which is fixed to 1000 in the ShinyApp.}

\item{post_iter}{Numeric.  The number of iterations to estimate the posterior distribution which is fixed to 1000 in the ShinyApp.}

\item{sigma}{Numeric. The standard deviation of the DV for the groups.}

\item{prior_location}{Numeric. The scale of the Cauchy prior which is fixed to 1 / sqrt(2) in the ShinyApp.}

\item{max_n}{Numeric. The maximum group size which is fixed to 500 in the ShinyApp.}

\item{seed}{Numeric. Seed for the random calculations.}

\item{n}{Numeric. The sample size per group during the tpr optimization process.}
}
\value{
The function returns a list of two named numeric vectors.
  The second `n1` the determined sample size per group.
  The third `tpr_out` is the TPR corresponding to the determined sample sizes.
}
\description{
Script for pre-calculation.
The present method provides an expected sample size such that
compelling evidence in the form of a Bayes factor can be collected
for a given eq band with a certain long-run probability.
}
