% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_rope_anova.R
\name{ssp_anova_rope}
\alias{ssp_anova_rope}
\title{Determine sample size with the Region of Practical Equivalence Method (ROPE)}
\usage{
ssp_anova_rope(
  mu,
  effect,
  eq_band,
  tpr,
  ci,
  prior_scale,
  iter,
  post_iter = 1000,
  sigma = 1,
  prior_location = 0,
  max_n = 10001,
  seed = NULL
)
}
\description{
Script for pre-calculations!
The present method provides an expected sample size such that
compelling evidence in the form of the Highest Density Interval can be collected
for a given eq band with a certain long-run probability.
}
