% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_bfda.R
\name{ssp_bfda}
\alias{ssp_bfda}
\title{Determine sample size with Bayes Factor Design Analysis (BFDA)}
\usage{
ssp_bfda(delta, thresh = 10, tpr = 0.8, n_rep = 1000)
}
\arguments{
\item{delta}{Numeric. the expected population effect size}

\item{thresh}{Integer. the Bayes factor threshold for inference}

\item{tpr}{Numeric. the long-run probability of obtaining a Bayes factor at least
as high as the critical threshold favoring superiority, given Delta}

\item{n_rep}{Integer. the number of simulations}
}
\value{
A list of two numeric values.
}
\description{
lorem ipsum
}
