% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_eq_bf.R
\name{ssp_eq_bf}
\alias{ssp_eq_bf}
\title{Determine sample size with Equivalence interval Bayes factor method}
\usage{
ssp_eq_bf(tpr, eq_band, delta, thresh = 10, tol = 1e-04, granularity = 300)
}
\arguments{
\item{tpr}{Numeric. The desired long-run probability of obtaining a Bayes factor at least as high as the Threshold, given Delta.}

\item{eq_band}{Numeric. The chosen width of the equivalence region.}

\item{delta}{Numeric. The expected population effect size.}

\item{thresh}{Integer. Critical threshold for the Bayes factor.}

\item{tol}{Numeric. TODO}

\item{granularity}{Numeric. TODO}
}
\value{
The function returns a list of two named numeric vectors.
The first `n1`  the determined sample size per group.
The third `npower` is the TPR corresponding to the determined sample sizes.
}
\description{
A Bayesian statistical testing approach aimed at establishing
equivalence between two groups.
}
\examples{
\dontrun{
SampleSizePlanner::ssp_eq_bf(tpr = 0.8, delta = 0, eq_band = 0.2, thresh = 10)
}
}
