% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_infer_bf.R
\name{ssp_infer_bf}
\alias{ssp_infer_bf}
\alias{infer_bf}
\title{Determine sample size with Non-inferiority Bayes factor method}
\usage{
ssp_infer_bf(
  tpr,
  ni_margin,
  delta,
  thresh = 10,
  tol = 1e-04,
  granularity = 300,
  prior_location = 0,
  prior_scale = 1/sqrt(2),
  max_n = 10001
)

infer_bf(
  n1,
  ni_margin,
  delta,
  thresh,
  tol,
  granularity,
  prior_location,
  prior_scale
)
}
\arguments{
\item{tpr}{Numeric. The desired long-run probability of obtaining a Bayes factor at least as high as the Threshold, given Delta.}

\item{ni_margin}{Numeric. A non-negative number representing the non-inferiority margin.}

\item{delta}{Numeric. The expected population effect size.}

\item{thresh}{Integer. Critical threshold for the Bayes factor.}

\item{tol}{Numeric. Relative accuracy requested.}

\item{granularity}{Numeric. Relative precision of the tpr estimates, higher values mean more precision.}

\item{prior_location}{Numeric. Location of the Cauchy prior distribution.}

\item{prior_scale}{Numeric. Scale of the Cauchy prior distribution.}

\item{max_n}{Integer. The maximum number of participants per group (both groups are assumed to have equal sample size).}
}
\value{
The function returns a list of two named numeric vectors.
The first `n1`  the determined sample size per group.
The third `tpr_out` is the TPR corresponding to the determined sample sizes.
}
\description{
A Bayesian statistical testing approach aimed at establishing
non-inferiority of one condition compared to the other.
}
\examples{
\dontrun{
SampleSizePlanner::ssp_infer_bf(tpr = 0.8, delta = 0, ni_margin = 0.2, thresh = 10)
}
}
