% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_power_traditional.R
\name{ssp_power_traditional}
\alias{ssp_power_traditional}
\title{Determine sample size with the Classical power analysis method}
\usage{
ssp_power_traditional(tpr, delta, max_n, alpha = 0.05)
}
\arguments{
\item{tpr}{Numeric. The desired long-run probability of obtaining a significant result with a one-sided t-test, given Delta.}

\item{delta}{Numeric. The expected population effect size.}

\item{max_n}{Integer. The maximum number of participants per group (both groups are assumed to have equal sample size).}

\item{alpha}{Numeric. The level of significance.}
}
\value{
The function returns a list of one named numeric vector.
The vector called `n1` contains the determined sample size per group
for the given design.
}
\description{
This method is used to estimate the minimum sample size that
a design needs to reach a statistical power, given a  desired
significance level and expected effect size.
}
\examples{
\dontrun{
SampleSizePlanner::ssp_power_traditional(tpr = 0.8, delta = 0.5, max_n = 5000, alpha = 0.05)
}
}
