% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr_optim.R
\name{tpr_optim}
\alias{tpr_optim}
\title{Optimize output TPR}
\usage{
tpr_optim(fun, range, delta, tpr, ...)
}
\arguments{
\item{fun}{Function. The sample size determining method to use.}

\item{range}{Integer. Range of sample sizes of interest.}

\item{delta}{Numeric. The expected population effect size.}

\item{tpr}{Numeric. The desired long-run probability.}

\item{...}{Additional arguments to pass to `fun`.}
}
\value{
The function returns a list of two. `n1` the resulting sample size and `tpr_out`
the associated true positive rate with that sample size.
}
\description{
The function optimizes the true positive rate for a range of sample sizes
for several methods in the package.
}
