% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ssp_bf_predetermined.R
\name{mod_ssp_bf_predetermined_ui}
\alias{mod_ssp_bf_predetermined_ui}
\alias{mod_ssp_bf_predetermined_server}
\title{mod_ssp_bf_thresh_ui and mod_ssp_bf_thresh_server}
\usage{
mod_ssp_bf_predetermined_ui(id)

mod_ssp_bf_predetermined_server(id)
}
\description{
A shiny Module.
}
\keyword{internal}
