% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ssp_bfda.R
\name{mod_ssp_bfda_ui}
\alias{mod_ssp_bfda_ui}
\alias{mod_ssp_bfda_server}
\title{mod_ssp_bfda_ui and mod_ssp_bfda_server}
\usage{
mod_ssp_bfda_ui(id)

mod_ssp_bfda_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
