% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ssp_tost.R
\name{mod_ssp_tost_ui}
\alias{mod_ssp_tost_ui}
\alias{mod_ssp_tost_server}
\title{mod_ssp_tost_ui and mod_ssp_tost_server}
\usage{
mod_ssp_tost_ui(id)

mod_ssp_tost_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
