% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_anova_bf.R, R/ssp_power_traditional_anova.R
\name{ssp_anova_bf}
\alias{ssp_anova_bf}
\alias{twoway_ANOVA_bf_pwr}
\alias{twowayANOVApwr}
\title{Determine sample size with Bayesian Anova method}
\usage{
ssp_anova_bf(
  effect,
  mu,
  tpr,
  thresh,
  prior_scale,
  iter,
  max_n = 10001,
  max_bf = 1e+08,
  sigma = 1,
  seed = NULL
)

twoway_ANOVA_bf_pwr(
  effect = effect,
  iter = iter,
  n = n1,
  mu = mu,
  sigma = sigma,
  thresh = thresh,
  max_bf = max_bf,
  prior_scale = prior_scale,
  seed = NULL
)

twowayANOVApwr(effect, iter, n1, mu, sigma, alpha, seed)
}
\arguments{
\item{effect}{Character. The effect of interest (main effect A, main effect B, interaction effect).}

\item{mu}{Numeric.  The mean of the DV for each group.}

\item{tpr}{Numeric. The long-run probability of obtaining a Bayes factor at least
as high as the critical threshold favoring superiority, given mu.}

\item{thresh}{Integer. The Bayes factor threshold for inference.}

\item{prior_scale}{Numeric. Scale of the Cauchy prior distribution.}

\item{iter}{Integer. The number of iterations.}

\item{max_n}{Integer. The maximum number of participants per group (all groups are assumed to have equal sample size).}

\item{max_bf}{Numeric. The maximum Bayes Factor value to reduce computation time. If all Bayes Factors in the first 10 iterations exceeded the max_bf,
we set the TPR equal to 1 and proceeded with the next sample size.}

\item{sigma}{Numeric. The standard deviation of the DV for the groups.}

\item{seed}{Numeric. The seed for reproducibility.}

\item{n}{Numeric. The sample size per group during the tpr optimization process.}

\item{n1}{Numeric. The sample size per group during the tpr optimization process.}
}
\value{
The function returns a list of four named numeric vectors.
The first `n1` is the range of determined sample sizes for the given design.
The second `tpr_out` is the range of TPRs that were provided as a parameter.
}
\description{
The present method provides an expected sample size such that
compelling evidence in the form of a Bayes factor can be collected
given the group means.
}
\examples{
\dontrun{
SampleSizePlanner::ssp_anova_bf(
  effect = "Interaction Effect", tpr = 0.8, max_n = 10001, mu = c(1.5, 1.5, 0, 1), sigma = 2,
  seed = NULL, thresh = 10, prior_scale = 1 / sqrt(2)
 )
}
}
