% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_bayesian_anova.R
\name{ssp_anova_bf}
\alias{ssp_anova_bf}
\title{Determine sample size with Bayes Factor Design Analysis (BFDA)}
\usage{
ssp_anova_bf(
  effect = c("Main Effect 1", "Main Effect 2", "Interaction Effect"),
  iter = 1000,
  max_n,
  mu,
  sigma,
  seed = NULL,
  tpr,
  thresh = 10,
  prior_scale = 1/sqrt(2)
)
}
\arguments{
\item{iter}{Integer. The number of simulations.}

\item{max_n}{Integer. The maximum number of participants per group (all groups are assumed to have equal sample size).}

\item{mu}{Numeric. The expected population mean values.}

\item{tpr}{Numeric. The long-run probability of obtaining a Bayes factor at least
as high as the critical threshold favoring superiority, given mu.}

\item{thresh}{Integer. The Bayes factor threshold for inference.}

\item{prior_scale}{Numeric. Scale of the Cauchy prior distribution.}
}
\value{
The function returns a list of four named numeric vectors.
The first `n1` is the range of determined sample sizes for the given design.
The second `tpr_out` is the range of TPRs that were provided as a parameter.
The third `effect` is the type of effect to evaluate, such as "Main Effect 1," "Main Effect 2," or "Interaction Effect."
}
\description{
The present method provides an expected sample size such that
compelling evidence in the form of a Bayes factor can be collected
for a given effect size with a certain long-run probability when
allowing for sequential testing.
}
\examples{
\dontrun{
SampleSizePlanner::ssp_anova_bf(
  tpr = 0.8, max_n = 400, mu = c(1, 1.2, 1.5, 1.3), sigma = 2,
  seed = NULL, thresh = 10, prior_scale = 1 / sqrt(2)
 )
}
}
