% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_eq_anova.R
\name{ssp_anova_eq}
\alias{ssp_anova_eq}
\title{Determine sample size with the Bayesian Equivalence Interval Method}
\usage{
ssp_anova_eq(
  mu,
  effect,
  eq_band,
  tpr,
  thresh,
  prior_scale,
  iter,
  post_iter = 1000,
  sigma = 1,
  prior_location = 0,
  max_n = 10001,
  seed = NULL
)
}
\description{
Script for pre-calculation.
The present method provides an expected sample size such that
compelling evidence in the form of a Bayes factor can be collected
for a given eq band with a certain long-run probability.
}
