% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_anova_rope.R
\name{ssp_anova_rope}
\alias{ssp_anova_rope}
\alias{twoway_ANOVA_rope_pwr}
\title{Determine sample size with the Region of Practical Equivalence Method (ROPE)}
\usage{
ssp_anova_rope(
  mu,
  effect,
  eq_band,
  tpr,
  ci,
  prior_scale,
  iter,
  post_iter = 1000,
  sigma = 1,
  prior_location = 0,
  max_n = 10001,
  seed = NULL
)

twoway_ANOVA_rope_pwr(
  n = n1,
  effect = effect,
  eq_band = eq_band,
  iter = iter,
  post_iter = post_iter,
  mu = mu,
  sigma = sigma,
  ci = ci,
  prior_scale = prior_scale,
  prior_location = prior_location,
  seed = NULL
)
}
\arguments{
\item{mu}{Numeric. The mean of the DV for each group.}

\item{effect}{Character. The effect of interest (main effect A, main effect B).}

\item{eq_band}{Numeric. The margin of the standardized ROPE interval.}

\item{tpr}{Numeric. The desired long-run probability of the HDI fully falling inside the ROPE, given the means.}

\item{ci}{Numeric. The percentage of the HDI which is fixed to 0.95 in the ShinyApp.}

\item{prior_scale}{Numeric. The scale of the Cauchy prior which is fixed to 1 / sqrt(2) in the ShinyApp.}

\item{iter}{Numeric. The number of iterations to calculate the TPR which is fixed to 1000 in the ShinyApp.}

\item{post_iter}{Numeric. The number of iterations to estimate the posterior distribution.}

\item{sigma}{Numeric. The standard deviation of the DV for the groups.}

\item{prior_location}{Numeric. The scale of the Cauchy prior which is fixed to 1 / sqrt(2) in the ShinyApp.}

\item{max_n}{Numeric. The maximum group size which is fixed to 500 in the ShinyApp.}

\item{seed}{Numeric. Seed for the random calculations.}
}
\value{
The function returns a list of two named numeric vectors.
  The second `n1` the determined sample size per group.
  The third `tpr_out` is the TPR corresponding to the determined sample sizes.
}
\description{
The present method provides an expected sample size such that
compelling evidence in the form of the Highest Density Interval can be collected
for a given eq band with a certain long-run probability.
}
