% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_bf_predetermined.R
\name{ssp_bf_predetermined}
\alias{ssp_bf_predetermined}
\title{Determine sample size with Predetermined sample size with Bayes factor method}
\usage{
ssp_bf_predetermined(
  tpr,
  delta,
  thresh = 10,
  max_n = 5000,
  tol = 1e-04,
  granularity = 300
)
}
\arguments{
\item{tpr}{Numeric. The long-run probability of obtaining a Bayes factor at least as high as the critical threshold favoring superiority, given Delta.}

\item{delta}{Numeric. The expected population effect size or a range of expected effect sizes.}

\item{thresh}{Integer. Critical threshold for the Bayes factor.}

\item{max_n}{Integer. The maximum number of participants per group (both groups are assumed to have equal sample size).}

\item{tol}{Numeric. TODO}

\item{granularity}{Numeric. TODO}
}
\value{
The function returns a list of three named numeric vectors.
The first `delta` is the range of deltas provided for the function.
The second `n1` the determined sample size per group.
The third `npower` is the TPR corresponding to the determined sample sizes
with the given delta.
}
\description{
The present method calculates the corresponding default Bayes factor for
a t-test statistic with Cauchy prior distribution centered on zero with
scale parameter 1/sqrt(2) for several sample sizes.
}
\examples{
\dontrun{
SampleSizePlanner::ssp_bf_predetermined(tpr = 0.8, delta = 0.5, thresh = 10, max_n = 5000)
}
}
