% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_bfda.R
\name{ssp_bfda}
\alias{ssp_bfda}
\title{Determine sample size with Bayes Factor Design Analysis (BFDA)}
\usage{
ssp_bfda(
  tpr = 0.8,
  delta,
  thresh = 10,
  n_rep = 1000,
  prior_scale = 1/sqrt(2),
  max_n = 1500
)
}
\arguments{
\item{tpr}{Numeric. The long-run probability of obtaining a Bayes factor at least
as high as the critical threshold favoring superiority, given Delta.}

\item{delta}{Numeric. The expected population effect size.}

\item{thresh}{Integer. The Bayes factor threshold for inference.}

\item{n_rep}{Integer. The number of simulations.}

\item{prior_scale}{Numeric. Scale of the Cauchy prior distribution.}

\item{max_n}{Integer. The maximum number of participants per group (both groups are assumed to have equal sample size).}
}
\value{
The function returns a list of four named numeric vectors.
The first `tpr` is the range of TPRs that were provided as a parameter.
The second `n1` is the range of determined sample sizes for the given design.
The third `h0` is the frequency of Bayes factor providing evidence with the
given threshold for the null hypothesis.
The fourth `ha` is the same as `h0` but for the alternative hypothesis.
}
\description{
The present method provides an expected sample size such that
compelling evidence in the form of a Bayes factor can be collected
for a given effect size with a certain long-run probability when
allowing for sequential testing.
}
\examples{
\dontrun{
SampleSizePlanner::ssp_bfda(tpr = 0.8, delta = 0.5, thresh = 10, n_rep = 1000)
}
}
