% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssp_power_curve.R
\name{ssp_power_curve}
\alias{ssp_power_curve}
\title{Calculating a power curve}
\usage{
ssp_power_curve(delta, tpr, max_n = 5000)
}
\arguments{
\item{delta}{Numeric. the range of hypothetical population effect sizes}

\item{tpr}{Numeric. the desired long run probability of obtaining a significant result with TOST, given Delta}

\item{animated}{Logical. if TRUE the output plot is animated}
}
\value{
The function returns a ggplot2 object.
}
\description{
lorem ipsum
}
